F1.OpenF1.DatasetBuilder/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openf1-dataset-builder"
version = "0.1.0"
description = "Pipeline OpenF1 para gerar datasets analiticos e publicar no MLflow"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.31",
  "pandas>=2.2",
  "pyarrow>=15.0",
  "PyYAML>=6.0",
  "mlflow>=2.12",
  "numpy>=1.26",
  "boto3>=1.34",
  "scikit-learn>=1.4",
  "fastapi>=0.110",
  "uvicorn>=0.27"
]

[tool.setuptools]
package-dir = {"" = "f1_dataset/src"}

[tool.setuptools.packages.find]
where = ["f1_dataset/src"]
include = [
  "config",
  "clients",
  "discovery",
  "collectors",
  "processors",
  "feature_engineering",
  "validators",
  "publishers",
  "orchestration",
  "jobs",
  "modeling",
  "api"
]
